+++ code/php/sdk/src/Defaults.php
@@ -5,7 +5,7 @@ use starfederation\datastar\enums\FragmentMergeMode;
 
 class Defaults
 {
-    public const DEFAULT_SETTLE_TIME = 300;
+    public const DEFAULT_SETTLE_DURATION = 300;
     public const DEFAULT_SSE_SEND_RETRY = 1000;
     public const DEFAULT_FRAGMENT_MERGE_MODE = FragmentMergeMode::Morph;
 }