+++ code/php/sdk/src/events/EventTrait.php
@@ -1,0 +1,12 @@
+<?php
+/**
+ * @copyright Copyright (c) PutYourLightsOn
+ */
+
+namespace starfederation\datastar\events;
+
+trait EventTrait
+{
+    public ?string $eventId = null;
+    public ?int $retryDuration = null;
+}