+++ code/ts/library/src/bundles/datastar.ts
@@ -1,62 +1,68 @@
-import { Datastar } from "library/src/engine";
-import { isFetchingActionPlugin } from "../plugins/actions/backend/isFetching";
-import { RemoteActionPlugin } from "../plugins/actions/backend/remote";
-import {
-    DeleteSSEActionPlugin,
-    GetSSEActionPlugin,
-    PatchSSEActionPlugin,
-    PostSSEActionPlugin,
-    PutSSEActionPlugin,
-} from "../plugins/actions/backend/sse";
-import { ClipboardActionPlugin } from "../plugins/actions/dom/clipboard";
-import { SetAllActionPlugin } from "../plugins/actions/logic/setAll";
-import { ToggleAllActionPlugin } from "../plugins/actions/logic/toggleAll";
-import { ClampFitActionPlugin } from "../plugins/actions/math/clampFit";
-import { ClampFitIntActionPlugin } from "../plugins/actions/math/clampFitInt";
-import { FitActionPlugin } from "../plugins/actions/math/fit";
-import { FitIntActionPlugin } from "../plugins/actions/math/fitInt";
-import { FetchIndicatorPlugin } from "../plugins/attributes/backend/fetchIndicator";
-import { HeaderPlugin } from "../plugins/attributes/backend/header";
-import { ReplaceUrlAttributePlugin } from "../plugins/attributes/backend/replaceUrl";
-import { BindAttributePlugin } from "../plugins/attributes/basics/bind";
-import { ClassAttributePlugin } from "../plugins/attributes/basics/class";
-import { ModelAttributePlugin } from "../plugins/attributes/basics/model";
-import { OnAttributePlugin } from "../plugins/attributes/basics/on";
-import { TextAttributePlugin } from "../plugins/attributes/basics/text";
-import { IntersectionAttributePlugin } from "../plugins/attributes/visibility/intersects";
-import { ScrollIntoViewAttributePlugin } from "../plugins/attributes/visibility/scrollIntoView";
-import { ShowAttributePlugin } from "../plugins/attributes/visibility/show";
-import { TeleportAttributePlugin } from "../plugins/attributes/visibility/teleport";
-import { ViewTransitionAttributePlugin } from "../plugins/attributes/visibility/viewTransition";
-import { PersistAttributePlugin } from "../plugins/storage/persist";
+import { Datastar } from "../engine";
+import { IsFetching } from "../plugins/actions/backend/isFetching";
+import { RemoteSignals } from "../plugins/actions/backend/remote";
+import { DeleteSSE } from "../plugins/actions/backend/sseDelete";
+import { GetSSE } from "../plugins/actions/backend/sseGet";
+import { PatchSSE } from "../plugins/actions/backend/ssePatch";
+import { PostSSE } from "../plugins/actions/backend/ssePost";
+import { PutSSE } from "../plugins/actions/backend/ssePut";
+import { Clipboard } from "../plugins/actions/dom/clipboard";
+import { SetAll } from "../plugins/actions/logic/setAll";
+import { ToggleAll } from "../plugins/actions/logic/toggleAll";
+import { ClampFit } from "../plugins/actions/math/clampFit";
+import { ClampFitInt } from "../plugins/actions/math/clampFitInt";
+import { Fit } from "../plugins/actions/math/fit";
+import { FitInt } from "../plugins/actions/math/fitInt";
+import { FetchIndicator } from "../plugins/attributes/backend/fetchIndicator";
+import { Header } from "../plugins/attributes/backend/header";
+import { ReplaceUrl } from "../plugins/attributes/backend/replaceUrl";
+import { Bind } from "../plugins/attributes/dom/bind";
+import { Class } from "../plugins/attributes/dom/class";
+import { Model } from "../plugins/attributes/dom/model";
+import { On } from "../plugins/attributes/dom/on";
+import { Text } from "../plugins/attributes/dom/text";
+import { Persist } from "../plugins/attributes/storage/persist";
+import { Intersection } from "../plugins/attributes/visibility/intersects";
+import { ScrollIntoView } from "../plugins/attributes/visibility/scrollIntoView";
+import { Show } from "../plugins/attributes/visibility/show";
+import { Teleport } from "../plugins/attributes/visibility/teleport";
+import { ViewTransition } from "../plugins/attributes/visibility/viewTransition";
 
 Datastar.load(
-    RemoteActionPlugin,
-    GetSSEActionPlugin,
-    PostSSEActionPlugin,
-    PutSSEActionPlugin,
-    PatchSSEActionPlugin,
-    DeleteSSEActionPlugin,
-    isFetchingActionPlugin,
-    ClipboardActionPlugin,
-    SetAllActionPlugin,
-    ToggleAllActionPlugin,
-    FitActionPlugin,
-    FitIntActionPlugin,
-    ClampFitActionPlugin,
-    ClampFitIntActionPlugin,
-    BindAttributePlugin,
-    ClassAttributePlugin,
-    ModelAttributePlugin,
-    OnAttributePlugin,
-    TextAttributePlugin,
-    FetchIndicatorPlugin,
-    HeaderPlugin,
-    ReplaceUrlAttributePlugin,
-    IntersectionAttributePlugin,
-    ScrollIntoViewAttributePlugin,
-    ShowAttributePlugin,
-    TeleportAttributePlugin,
-    ViewTransitionAttributePlugin,
-    PersistAttributePlugin,
+    // actions/backend
+    IsFetching,
+    RemoteSignals,
+    DeleteSSE,
+    GetSSE,
+    PatchSSE,
+    PostSSE,
+    PutSSE,
+    // actions/dom
+    Clipboard,
+    // actions/logic
+    SetAll,
+    ToggleAll,
+    // actions/math
+    ClampFit,
+    ClampFitInt,
+    Fit,
+    FitInt,
+    // attributes/backend
+    FetchIndicator,
+    Header,
+    ReplaceUrl,
+    // attributes/basics
+    Bind,
+    Class,
+    Model,
+    On,
+    Text,
+    // attributes/storage
+    Persist,
+    // attributes/visibility
+    Intersection,
+    ScrollIntoView,
+    Show,
+    Teleport,
+    ViewTransition,
 );