+++ code/ts/library/src/plugins/actions/math/clampFitInt.ts
@@ -1,6 +1,11 @@
-import { ActionPlugin, AttributeContext } from "library/src/engine";
+// Authors: Delaney Gillilan
+// Icon: material-symbols:fit-screen
+// Slug: Clamp a value to a new range and round to the nearest integer
+// Description: This action clamps a value to a new range. The value is first scaled to the new range, then clamped to the new range. This is useful for scaling a value to a new range, then clamping it to that range. The result is then rounded to the nearest integer.
 
-export const ClampFitIntActionPlugin: ActionPlugin = {
+import { ActionPlugin, AttributeContext } from "../../../engine";
+
+export const ClampFitInt: ActionPlugin = {
     pluginType: "action",
     name: "clampFitInt",
     method: (