+++ code/ts/library/src/plugins/actions/math/fit.ts
@@ -1,6 +1,11 @@
-import { ActionPlugin, AttributeContext } from "library/src/engine";
+// Authors: Delaney Gillilan
+// Icon: material-symbols-light:fit-width
+// Slug: Linearly fit a value to a new range
+// Description: This action linearly fits a value to a new range. The value is first scaled to the new range.  Note it is not clamped to the new range.
 
-export const FitActionPlugin: ActionPlugin = {
+import { ActionPlugin, AttributeContext } from "../../../engine";
+
+export const Fit: ActionPlugin = {
     pluginType: "action",
     name: "fit",
     method: (