+++ code/ts/library/src/plugins/attributes/storage/persist.ts
@@ -1,12 +1,17 @@
+// Authors: Delaney Gillilan
+// Icon: mdi:floppy-variant
+// Slug: Persist data to local storage or session storage
+// Description: This plugin allows you to persist data to local storage or session storage.  Once you add this attribute the data will be persisted to local storage or session storage.
+
 import {
     AttributePlugin,
     DATASTAR,
     DATASTAR_EVENT,
     DatastarEvent,
-} from "library/src/engine";
-import { remoteSignals } from "library/src/utils/signals";
+} from "../../../engine";
+import { remoteSignals } from "../../../utils/signals";
 
-export const PersistAttributePlugin: AttributePlugin = {
+export const Persist: AttributePlugin = {
     pluginType: "attribute",
     prefix: "persist",
     allowedModifiers: new Set(["local", "session", "remote"]),
@@ -36,6 +41,7 @@ export const PersistAttributePlugin: AttributePlugin = {
             if (keys.size > 0) {
                 const newStore: Record<string, any> = {};
                 for (const key of keys) {
+                    const parts = key.split(".");
                     let newSubstore = newStore;
                     let subStore = store;
                     for (let i = 0; i < parts.length - 1; i++) {
@@ -70,7 +76,7 @@ export const PersistAttributePlugin: AttributePlugin = {
 
         window.addEventListener(DATASTAR_EVENT, storeUpdateHandler);
 
-        let marshalledStore: string | null = null;
+        let marshalledStore: string | null;
 
         if (storageType === "session") {
             marshalledStore = window.sessionStorage.getItem(key);