+++ code/ts/library/src/plugins/preprocessors/core/actions.ts
@@ -1,8 +1,8 @@
-import { PreprocessorPlugin, RegexpGroups } from "library/src/engine";
-import { wholePrefixSuffix } from "library/src/utils/regex";
+import { PreprocessorPlugin, RegexpGroups } from "../../../engine";
+import { wholePrefixSuffix } from "../../../utils/regex";
 
 // Replacing $$action(args) with ctx.actions.action(ctx, args)
-export const ActionsProcessorPlugin: PreprocessorPlugin = {
+export const ActionsProcessor: PreprocessorPlugin = {
     name: "action",
     pluginType: "preprocessor",
     regexp: wholePrefixSuffix("$\\$", "action", "(?<call>\\((?<args>.*)\\))?"),